hehehe

+++ Lone Wiz/Assets/Scripts/Abilities Scripts/Upgrade.cs
@@ -4,9 +4,9 @@ public class Upgrade : MonoBehaviour
 {
     public string AbilityName { get; set; }
     public string UpgradeType { get; set; }
-    public int Change { get; set; }
+    public float Change { get; set; }
     public int Cost { get; set; }
-    public Upgrade(string name, string type, int change, int cost)
+    public Upgrade(string name, string type, float change, int cost)
     {
         AbilityName = name;
         UpgradeType = type;
